Working all-stocks view with TSE data
