+++ src/app/all-stocks/all-stocks.component.html
@@ -1,53 +1,70 @@
 <div>
-    <table mat-table [dataSource]="marketHoldings" class="mat-elevation-z8"
-        matSort matSortActive="created"
-        matSortDisableClear
-        matSortDirection="asc">
 
-      <!-- ticker Column -->
-      <ng-container matColumnDef="ticker">
-        <th mat-header-cell *matHeaderCellDef mat-sort-header>Ticker</th>
-        <td mat-cell *matCellDef="let record"> {{record.ticker}} </td>
-      </ng-container>
+  <mat-form-field>
+    <input matInput (keyup)="applyFilter($event.target.value)" placeholder="Filter">
+  </mat-form-field>
+  
+  <button class="mat-stroked-button" (click)="refreshMarketHoldings()">Refresh Holdings</button>
 
-      <!-- name Column -->
-      <ng-container matColumnDef="name">
-        <th mat-header-cell *matHeaderCellDef mat-sort-header>Name</th>
-        <td mat-cell *matCellDef="let record"> {{record.name}} </td>
-      </ng-container>
+  <mat-paginator 
+    [pageSizeOptions]="mpPageSizeOptions"
+    (page)="pageEvent = $event">
+  </mat-paginator> 
 
-      <!-- price Column -->
-      <ng-container matColumnDef="price">
-        <th mat-header-cell *matHeaderCellDef mat-sort-header>Price</th>
-        <td mat-cell *matCellDef="let record"> {{record.price}} </td>
-      </ng-container>
-      
-      <!-- exchange Column -->
-      <ng-container matColumnDef="exchange">
-        <th mat-header-cell *matHeaderCellDef mat-sort-header>Exchange</th>
-        <td mat-cell *matCellDef="let record"> {{record.exchange}} </td>
-      </ng-container>
+  <table mat-table [dataSource]="marketHoldings" class="mat-elevation-z8"
+      matSort matSortActive="created"
+      matSortDisableClear
+      matSortDirection="asc">
 
-      <!-- sector Column -->
-      <ng-container matColumnDef="sector">
-        <th mat-header-cell *matHeaderCellDef mat-sort-header>Sector</th>
-        <td mat-cell *matCellDef="let record"> {{record.sector}} </td>
-      </ng-container>
+    <!-- ticker Column -->
+    <ng-container matColumnDef="ticker">
+      <th mat-header-cell *matHeaderCellDef mat-sort-header>Ticker</th>
+      <td mat-cell *matCellDef="let record"> {{record.ticker}} </td>
+    </ng-container>
 
-      <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-      <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
-    </table>
+    <!-- name Column -->
+    <ng-container matColumnDef="name">
+      <th mat-header-cell *matHeaderCellDef mat-sort-header>Name</th>
+      <td mat-cell *matCellDef="let record"> {{record.name}} </td>
+    </ng-container>
 
-    <mat-form-field>
-      <input matInput (keyup)="applyFilter($event.target.value)" placeholder="Filter">
-    </mat-form-field>
-    
-    <button class="mat-stroked-button" (click)="refreshMarketHoldings()">Refresh Holdings</button>
+    <!-- price Column -->
+    <ng-container matColumnDef="price">
+      <th mat-header-cell *matHeaderCellDef mat-sort-header>Price</th>
+      <td mat-cell *matCellDef="let record"> {{record.price}} </td>
+    </ng-container>
 
-    <mat-paginator 
-      [length]="marketHoldings.data.length"
-      [pageSize]="50"
-      [pageSizeOptions]="mpPageSizeOptions"
-      (page)="pageEvent = $event">
-    </mat-paginator>  
+    <!-- high52 Column -->
+    <ng-container matColumnDef="high52">
+      <th mat-header-cell *matHeaderCellDef mat-sort-header>52-week High</th>
+      <td mat-cell *matCellDef="let record"> {{record.high52}} </td>
+    </ng-container>
+
+    <!-- low52 Column -->
+    <ng-container matColumnDef="low52">
+      <th mat-header-cell *matHeaderCellDef mat-sort-header>52-week Low</th>
+      <td mat-cell *matCellDef="let record"> {{record.low52}} </td>
+    </ng-container>
+
+    <!-- marketcap Column -->
+    <ng-container matColumnDef="marketcap">
+      <th mat-header-cell *matHeaderCellDef mat-sort-header>Market Cap</th>
+      <td mat-cell *matCellDef="let record"> {{record.marketcap}} </td>
+    </ng-container>
+
+    <!-- exchange Column -->
+    <ng-container matColumnDef="exchange">
+      <th mat-header-cell *matHeaderCellDef mat-sort-header>Exchange</th>
+      <td mat-cell *matCellDef="let record"> {{record.exchange}} </td>
+    </ng-container>
+
+    <!-- sector Column -->
+    <ng-container matColumnDef="sector">
+      <th mat-header-cell *matHeaderCellDef mat-sort-header>Sector</th>
+      <td mat-cell *matCellDef="let record"> {{record.sector}} </td>
+    </ng-container>
+
+    <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+    <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
+  </table> 
 </div>
