+++ src/app/all-stocks/all-stocks.component.ts
@@ -1,6 +1,6 @@
-import { MarketDataStore } from '../stores/Market.store';
 import { MarketHolding } from '../models/holding.model';
-import { Component, OnInit, ViewChild } from '@angular/core';
+import { DataStore } from '../stores/data.store';
+import { Component, OnInit, ViewChild, AfterViewInit } from '@angular/core';
 import { MatTableDataSource } from '@angular/material/table';
 import { MatPaginator, PageEvent } from '@angular/material/paginator';
 import { MatSort } from '@angular/material/sort';
@@ -10,35 +10,27 @@ import { MatSort } from '@angular/material/sort';
   templateUrl: './all-stocks.component.html',
   styleUrls: ['./all-stocks.component.scss']
 })
-export class AllStocksComponent implements OnInit {
+export class AllStocksComponent implements OnInit, AfterViewInit {
 
-  marketHoldings = new MatTableDataSource<MarketHolding>();
+  marketHoldings: MatTableDataSource<MarketHolding> = new MatTableDataSource<MarketHolding>();
+  marketData: any;
+
   holdingsObjName = 'MarketHoldings';
 
-  displayedColumns: string[] = ['ticker', 'name', 'price', 'high52', 'low52', 'exchange', 'sector'];
-
-  @ViewChild(MatPaginator) paginator: MatPaginator;
-  mpPageSizeOptions: number[] = [10, 25, 50, 100];
-  pageEvent: PageEvent;
+  displayedColumns: string[] = ['ticker', 'name', 'price', 'high52', 'low52', 'marketcap', 'exchange', 'sector'];
 
   @ViewChild(MatSort) sort: MatSort;
+  @ViewChild(MatPaginator) paginator: MatPaginator;
+  mpPageSizeOptions: number[] = [25, 50, 100];
+  pageEvent: PageEvent;
 
-  constructor(public dataStore: MarketDataStore) {
 
+  constructor(public dataStore: DataStore) {
     this.dataStore.marketHoldingsUpdated.subscribe(
       (newData: any) => {
-        this.marketHoldings = new MatTableDataSource(newData);
-        this.marketHoldings.paginator = this.paginator;
-        this.marketHoldings.sort = this.sort;
+        this.marketData = newData;
       }
     );
-
-    this.marketHoldings.sortingDataAccessor = (item, property) => {
-      switch (property) {
-        default: return item[property];
-      }
-    };
-
   }
 
   ngOnInit() {
@@ -46,6 +38,14 @@ export class AllStocksComponent implements OnInit {
       // use the local storage if there until HTTP call retrieves something
       JSON.parse(localStorage[this.dataStore.dataObjects.getCacheName(this.holdingsObjName)] || '[]')
     );
+  }
+
+  ngAfterViewInit() {
+    this.marketHoldings.sort = this.sort;
+    this.marketHoldings.paginator = this.paginator;
+    // TODO: browser error when roendering this way, but soooo much faster... look into this.
+    // functionality is great so not urgent
+    this.marketHoldings.data = this.marketData;
   }
 
   refreshMarketHoldings() {
