+++ src/app/stores/data.service.ts
@@ -6,37 +6,37 @@ import { catchError, tap, map } from 'rxjs/operators';
 
 @Injectable()
 export class DataService {
-    protected myHoldingsUrl = AppConfig.settings.my_holdings_sheet_url;
-    protected marketHoldingsUrl = AppConfig.settings.market_holdings_sheet_url;
+  protected myHoldingsUrl = AppConfig.settings.my_holdings_sheet_url;
+  protected tseHoldingsUrl = AppConfig.settings.market_holdings_sheet_url;
 
-    constructor(private httpClient: HttpClient) {}
+  constructor(private httpClient: HttpClient) {}
 
-    private handleError() {
-        return (error: any): Observable<any> => {
-          console.error(error);
-          return;
-        };
-      }
+  private handleError() {
+    return (error: any): Observable<any> => {
+      console.error(error);
+      return;
+    };
+  }
 
-    getMyHoldings(): Observable<any> {
-        console.log(`GET: ${this.myHoldingsUrl}`);
+  getMyHoldings(): Observable<any> {
+    console.log(`GET: MyHoldings`);
 
-        const res = this.httpClient.get<any>(this.myHoldingsUrl).pipe(
-          map(obj => obj.feed.entry),
-          catchError(this.handleError())
-        );
+    const res = this.httpClient.get<any>(this.myHoldingsUrl).pipe(
+      map(obj => obj.feed.entry),
+      catchError(this.handleError())
+    );
 
-        return res;
-    }
+    return res;
+  }
 
-    getMarketHoldings(): Observable<any> {
-      console.log(`GET: ${this.marketHoldingsUrl}`);
+  getTSEHoldings(): Observable<any> {
+    console.log(`GET: MarketHoldings`);
 
-      const res = this.httpClient.get<any>(this.marketHoldingsUrl).pipe(
-        map(obj => obj.feed.entry),
-        catchError(this.handleError())
-      );
+    const res = this.httpClient.get<any>(this.tseHoldingsUrl).pipe(
+      map(obj => obj.feed.entry),
+      catchError(this.handleError())
+    );
 
-      return res;
+    return res;
   }
 }
